fix: whole-codebase RCA + review sweep, plus config template MCMC cleanup
Fixes confirmed real and locally scoped across three parallel efforts this
session (deep-RCA debugging sweep, PR-review-toolkit sweep, and config
template cleanup):

- config/parameter_manager.py: guard against a config whose analysis_mode
  was overridden (e.g. CLI --mode) without updating
  initial_parameters.parameter_names to match; falls back to mode defaults
  instead of handing a wrong-length parameter vector downstream.
- config/{parameter_manager,heterodyne_parameter_manager}.py + templates:
  drop the vestigial "TruncatedNormal" BoundDict.type field and MCMC-only
  YAML keys -- Bayesian sampling is permanently out of scope for xpcsjax
  (see CLAUDE.md).
- core/diagonal_correction.py: force float64 on C2 input so an int-dtype
  HDF5 array doesn't truncate the corrected diagonal on write-back.
- core/physics.py: log real validation failures instead of silently
  swallowing them alongside genuine JAX tracer skips.
- core/physics_utils.py: NaN-preserving floor in
  calculate_diffusion_coefficient/calculate_shear_rate -- `x > 1e-10` is
  False for NaN, which was laundering a divergent parameter into a
  plausible finite value instead of propagating it.
- optimization/nlsq/wrapper.py: fix a wrong get_initial_values() call on
  the homodyne ParameterManager (that method only exists on the unrelated
  heterodyne class of the same name); fall back to the bounds midpoint.
- optimization/nlsq/parallel_accumulator.py: guard sigma<=0 in the
  out-of-core chunk residual/chi2 (1/sigma), mirroring
  heterodyne_stratified_data.py -- one degenerate chunk previously poisoned
  the whole OOC accumulation with inf/nan.
- gui/ipc/worker.py: surface a plot-generation failure as a LogLine instead
  of silently reporting a plain Finished with plots quietly missing.
- gui/ipc/emitter.py: exempt WARNING/ERROR/CRITICAL LogLines from the 20 Hz
  telemetry throttle so a fast-iterating fit can't drop the log lines
  status_manager's severity coloring exists to surface.
- service/config.py: drop an AttributeError guard around
  _normalize_analysis_mode() that could never legitimately fire and would
  otherwise mask a real bug inside the method's own call chain.
- viz/nlsq_plots.py: count only angles whose evaluation actually raised
  toward n_failed_angles, not any angle that merely contains a NaN value
  (degenerate physics can legitimately produce partial NaNs without a
  failure).

Full suite (tests/data, config, integration, parity, service, cli, gui,
core, optimization) passes 2198/2198.

+++ xpcsjax/core/diagonal_correction.py
@@ -339,7 +339,9 @@ def _diagonal_correction_numpy(
 ) -> np.ndarray:
     """NumPy implementation of diagonal correction with multiple methods."""
     config = config or {}
-    c2_np = np.asarray(c2_mat)
+    # dtype=float64: an integer-dtype input (raw HDF5 C2 can be int-typed)
+    # would otherwise truncate the corrected diagonal on write-back below.
+    c2_np = np.asarray(c2_mat, dtype=np.float64)
 
     if method == "basic":
         return _basic_correction_numpy(c2_np)
@@ -359,7 +361,8 @@ def _diagonal_correction_batch_numpy(
 ) -> np.ndarray:
     """Batch NumPy implementation with pre-allocated arrays."""
     config = config or {}
-    c2_np = np.asarray(c2_matrices)
+    # dtype=float64: see _diagonal_correction_numpy's identical guard.
+    c2_np = np.asarray(c2_matrices, dtype=np.float64)
     n_phi = c2_np.shape[0]
     size = c2_np.shape[1]
